Babel 7, tree shaking, replace latest with env
