+++ .babelrc
@@ -1,33 +1,4 @@
 {
-  "presets": [
-    [
-      "latest",
-      {
-        "loose": true,
-        "modules": false
-      }
-    ],
-    "stage-0",
-    "stage-3",
-    "react"
-  ],
-  "plugins": [
-    "universal-import",
-    "react-hot-loader/babel",
-    "transform-class-properties",
-    [
-      "transform-runtime",
-      {
-        "helpers": false,
-        "polyfill": false,
-        "regenerator": true,
-        "moduleName": "babel-runtime"
-      }
-    ]
-  ],
-  "env": {
-    "test": {
-      "presets": [["latest", { "modules": true }], "stage-0", "stage-3", "react"]
-    }
-  }
+  "presets": ["./babel-preset.js"],
+  "ignore": ["app.static.js"]
 }
