+++ package.json
@@ -17,35 +17,36 @@
     "prepublishOnly": "yarn build",
     "postpublish": "git push --tags",
     "test": "eslint src examples test && yarn run test-sample-build",
-    "test-sample-build": "yarn build && (cd test/minimal-website && ../../bin/react-static build)",
-    "test-webpack": "jest"
+    "test-sample-build": "yarn build && (cd test/minimal-website && ../../bin/react-static build)"
   },
   "repository": "nozzle/react-static",
   "devDependencies": {
-    "eslint": "^4.3.0",
-    "eslint-config-react-tools": "^1.0.6",
-    "jest": "^21.2.1",
+    "babel-eslint": "^8.2.1",
+    "eslint": "^4.16.0",
+    "eslint-config-react-tools": "^1.1.1",
+    "eslint-plugin-class-property": "^1.1.0",
+    "eslint-plugin-flowtype": "^2.42.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.5.1",
     "rimraf": "^2.6.2"
   },
   "dependencies": {
+    "@babel/cli": "^7.0.0-beta.38",
+    "@babel/core": "^7.0.0-beta.38",
+    "@babel/plugin-transform-runtime": "^7.0.0-beta.38",
+    "@babel/preset-env": "^7.0.0-beta.38",
+    "@babel/preset-react": "^7.0.0-beta.38",
+    "@babel/preset-stage-0": "^7.0.0-beta.38",
+    "@babel/register": "^7.0.0-beta.38",
+    "@babel/runtime": "^7.0.0-beta.38",
+    "@std/esm": "^0.19.7",
     "@types/react": "^16.0.18",
     "@types/react-helmet": "^5.0.3",
     "@types/react-router-dom": "^4.2.0",
     "autoprefixer": "^7.1.5",
     "axios": "^0.16.2",
-    "babel-cli": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-universal-import": "^1.4.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-latest": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
-    "babel-runtime": "^6.26.0",
+    "babel-loader": "^8.0.0-beta.0",
     "case-sensitive-paths-webpack-plugin": "^2.1.1",
     "chalk": "^2.3.0",
     "circular-dependency-plugin": "^4.2.1",
@@ -72,7 +73,7 @@
     "react-dev-utils": "^4.0.1",
     "react-dom": "^16",
     "react-helmet": "^5.2.0",
-    "react-hot-loader": "^3.0.0-beta.7",
+    "react-hot-loader": "^3.1.3",
     "react-router-dom": "^4.2.2",
     "react-universal-component": "^2.8.1",
     "shorthash": "^0.0.2",
